\name{fragmentRthepackage-package}
\alias{fragmentRthepackage-package}
\alias{fragmentRthepackage}
\docType{package}
\title{
\packageTitle{fragmentRthepackage}
}
\description{
\packageDescription{fragmentRthepackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fragmentRthepackage}
\packageIndices{fragmentRthepackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fragmentRthepackage}

Maintainer: \packageMaintainer{fragmentRthepackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
