% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_fragment.R
\name{image_fragment}
\alias{image_fragment}
\title{Add a Graded Fragmentation Effect to a Magick Image}
\usage{
image_fragment(image, pixel_size, levels, pixel_colour = "white")
}
\arguments{
\item{image}{a magick image object}

\item{pixel_size}{a double representing the height and width of each image fragment}

\item{levels}{the number of fragmentation levels to generate, inclusive of the fully revealed image}

\item{pixel_colour}{a valid magick colour string such as "navyblue" or "#000000" that determines the fill colour of the obscured image fragments}
}
\description{
This function creates a set number of copies of a magick image, with an increasing proportion of random non-background "fragments" of the image obscured by coloured squares.
}
\examples{
stimulus <- image_read_svg("http://jeroen.github.io/images/tiger.svg", width = 400)

fragments <- image_fragment(stimulus, pixel_size = 20, levels = 10)
}
\keyword{fragmentation}
